moved the models out of the algorithms files
