+++ sdk/keyvault/keyvault-keys/src/index.ts
@@ -100,9 +100,9 @@ import {
   WrapResult,
   EncryptResult
 } from "./cryptographyClientModels";
-import { LocalSupportedAlgorithmName } from "./localCryptography/algorithms";
 
 import { parseKeyVaultKeyId, KeyVaultKeyId } from "./identifier";
+import { LocalSupportedAlgorithmName } from "./localCryptography/models";
 
 export {
   CryptographyClientOptions,
