+++ sdk/keyvault/keyvault-keys/src/localCryptography/runOperation.ts
@@ -1,13 +1,13 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { LocalCryptographyUnsupportedError } from "./models";
 import {
-  localSupportedAlgorithms,
+  LocalCryptographyOperationName,
+  LocalCryptographyUnsupportedError,
   LocalSupportedAlgorithm,
-  LocalSupportedAlgorithmName,
-  LocalCryptographyOperationName
-} from "./algorithms";
+  LocalSupportedAlgorithmName
+} from "./models";
+import { localSupportedAlgorithms } from "./algorithms";
 import { JsonWebKey } from "../keysModels";
 import { convertJWKtoPEM } from "./conversions";
